Optimization Utilities

Newton

Utility functions for working with Newton's method.

gps.Newton.newton — Method

newton(f, x; ∇f=f', δ=eps(f(x)))

Find a root of f with derivative ∇f within accuracy of δ and start point x.

By default δ is the machine epsilon of the output type at 1 of f and ∇f is the automatic differentiation of f. These assume that f is a real-valued function.

Least Squares

A method to compute the least-squares optimization problem

gps.LeastSquares.leastsquares — Method

leastsquares(F, x₀; [δ])

Minimize F as much as possible within accuracy of δ and start point x₀.

F must be a vector-valued function of a vector-valued input. If F is a vector-valued function with a scalar input, call leastsquares(x->F(x[1]), [x₀]) instead.

This also assumes that F is written 100% in Julia and accepts any Real input. If this is not so, the automatic differentiation may not succeed.