Prelude

Functions visible to the entire gps project

gps.cart2ll — Function

cart2ll(coords, t=0)

Convert Cartesian coordinates into latitude, longitude and height

The inverse of ll2cart

gps.dms2rad — Function

dms2rad(d, m, s)

Given degrees, minutes and seconds of an angle convert to a radian value.

The inverse of rad2dms.

gps.ll2cart — Function

ll2cart(ψ, λ, h, t=0) Convert latitude, longitude, height and time to (x,y,z) coordinates

gps.rad2dms — Method

rad2dms(α)

Given a radian value convert to degrees, minutes and seconds.

The inverse of dms2rad.

gps.rot_z — Method

rot_z(α)

Generates a 3d "roll" rotation matrix about the z axis by α radians

gps.validatecoords — Method

validatecoords(x...)

Validate all coordinates as being elements of ℝ^3